+++ commanderbot/ext/automod/triggers/thread_updated.py
@@ -7,11 +7,11 @@ from commanderbot.ext.automod.triggers.abc.thread_base import ThreadBase
 
 
 @dataclass
-class ThreadCreated(ThreadBase):
+class ThreadUpdated(ThreadBase):
     """
-    Fires when an `on_thread_join` event is received without already being a member.
+    Fires when an `on_thread_update` event is received.
 
-    See: https://discordpy.readthedocs.io/en/master/api.html#discord.on_thread_join
+    See: https://discordpy.readthedocs.io/en/master/api.html#discord.on_thread_update
 
     Attributes
     ----------
@@ -19,8 +19,8 @@ class ThreadCreated(ThreadBase):
         The parent channels to match against. If empty, all channels will match.
     """
 
-    event_types = (events.ThreadCreated,)
+    event_types = (events.ThreadUpdated,)
 
 
 def create_trigger(data: Any) -> Trigger:
-    return ThreadCreated.from_data(data)
+    return ThreadUpdated.from_data(data)