+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `automod`:
+  - Added new thread fields to event metadata
+  - Implemented new `thread_updated` trigger
+  - Implemented new `thread_auto_archive_duration` condition
+
 ## [0.18.0] - 2021-10-04
 
 ### Added